Rating displays till one decimal point
ISSUE ID:ACUMOS-2612
Description:IST: Rating is not displaying till one decimal point

+++ acumos-portal-be/src/main/java/org/acumos/portal/be/util/PortalUtils.java
@@ -149,7 +149,7 @@ public class PortalUtils {
 		mlSolution.setDownloadCount(mlpSolution.getDownloadCount().intValue());
 		mlSolution.setRatingCount(mlpSolution.getRatingCount().intValue());
 		mlSolution.setViewCount(mlpSolution.getViewCount().intValue());
-		mlSolution.setSolutionRatingAvg(mlpSolution.getRatingAverageTenths() / 10);
+		mlSolution.setSolutionRatingAvg(mlpSolution.getRatingAverageTenths().floatValue() / 10);
 		mlSolution.setLastDownload(mlpSolution.getLastDownload());
 		if (mlpSolution.isFeatured() != null) {
 			mlSolution.setFeatured(mlpSolution.isFeatured());
